+++ CyReP/tcps/cborhelper.h
@@ -26,7 +26,7 @@ CborError
 cbor_value_ref_byte_string(
     CborValue *Cborstring,
     const uint8_t **Bstr,
-    uint32_t *BstrSize,
+    size_t *BstrSize,
     CborValue *Next
 );
 
@@ -34,7 +34,7 @@ CborError
 cbor_value_ref_text_string(
     CborValue *Cborstring,
     const uint8_t **Bstr,
-    uint32_t *BstrSize,
+    size_t *BstrSize,
     CborValue *Next
 );
